+++ optaPlanner-DB/src/main/java/com/scheduleplanner/optaPlannerDB/Service/interfaces/LessonService.java
@@ -13,6 +13,6 @@ public interface LessonService {
 
     Lesson updateLesson(Lesson l);
 
-    Boolean deleteLesson(Long id);
+    Boolean deleteLesson(int id);
 
 }
