+++ optaPlanner-DB/src/main/java/com/scheduleplanner/optaPlannerDB/Service/interfaces/RoomService.java
@@ -13,5 +13,5 @@ public interface RoomService {
 
     Room updateRoom(Room r);
 
-    Boolean deleteRoom(Long id);
+    Boolean deleteRoom(int id);
 }
